Templatize SQLHashStorage and SQLMultiStorage
