+++ src/game/DBScripts/ScriptMgr.cpp
@@ -2499,7 +2499,7 @@ int32 ScriptMgr::GetRandomRelayDbscriptFromTemplate(uint32 id)
 void ScriptMgr::CollectPossibleEventIds(std::set<uint32>& eventIds)
 {
     // Load all possible script entries from gameobjects
-    for (SQLStorageBase::SQLSIterator<GameObjectInfo> itr = sGOStorage.getDataBegin<GameObjectInfo>(); itr < sGOStorage.getDataEnd<GameObjectInfo>(); ++itr)
+    for (auto itr = sGOStorage.getDataBegin(); itr < sGOStorage.getDataEnd(); ++itr)
     {
         switch (itr->type)
         {
